+++ src/components/Spotify3DOptionsPanel.vue
@@ -296,6 +296,25 @@
           </div>
           <div class="field is-horizontal" v-if="options.base.hasKeychainAttachment">
             <div class="field-label is-small">
+              <label class="label">{{$t('keychainHoleDiameter')}}</label>
+            </div>
+            <div class="field-body">
+              <div class="field has-addons">
+                <div class="control">
+                  <input
+                    class="input is-small"
+                    type="number"
+                    v-model.number="options.base.keychainHoleDiameter"
+                  />
+                </div>
+                <p class="control">
+                  <a class="button is-static is-small">{{unit}}</a>
+                </p>
+              </div>
+            </div>
+          </div>
+          <div class="field is-horizontal" v-if="options.base.hasKeychainAttachment">
+            <div class="field-label is-small">
               <label class="label">{{$t('mirrorHoles')}}</label>
             </div>
             <div class="field-body">